Resolve contact form security findings
