+++ contact/index.html
@@ -23,5 +23,5 @@
 <section class="dark-section"><div class="section-block"><p class="eyebrow">What happens next</p><h2>A Simple, Structured Process</h2><div class="steps"><div class="step-card"><span>1</span><h3>Enquire</h3><p>Send your requirements through the form or book a discovery call.</p></div><div class="step-card"><span>2</span><h3>Understand</h3><p>We clarify your goals, current situation, scope and priorities.</p></div><div class="step-card"><span>3</span><h3>Recommend</h3><p>We identify the most practical solution and explain the approach.</p></div><div class="step-card"><span>4</span><h3>Scope &amp; Quote</h3><p>You receive an agreed scope, deliverables and commercial proposal where applicable.</p></div></div></div></section>
 <section><div class="section-block"><h2>Frequently Asked Questions</h2><div class="features"><div class="feature-card"><h3>Do I need to know exactly what service I need?</h3><p>No. Tell us what is not working or what you want to achieve. We can help identify the appropriate solution.</p></div><div class="feature-card"><h3>Do you work with NGOs?</h3><p>Yes. Project2Pixel supports NGOs and community organisations with digital infrastructure, communication, operational systems and practical implementation.</p></div><div class="feature-card"><h3>Can you work with an existing system?</h3><p>Where appropriate, yes. We can assess existing digital tools and workflows and identify improvements rather than replacing everything unnecessarily.</p></div><div class="feature-card"><h3>Do you offer ongoing support?</h3><p>Ongoing support can be scoped according to the organisation's requirements and project capacity.</p></div></div></div></section>
 </main>
-<footer><div class="footer-grid"><div><img src="/project2pixel-logo-480.png" alt="Project2Pixel Logo" class="footer-logo" width="480" height="480"><p>Project2Pixel provides practical business support and digital solutions for growing organisations.</p></div><div><h2 class="footer-heading">Navigate</h2><p><a href="/">Home</a> · <a href="/services/">Services</a> · <a href="/about/">About</a> · <a href="/contact/">Contact</a></p><p><a href="/privacy.html">Privacy Policy</a> · <a href="/terms.html">Terms of Service</a> · <a href="/cancellation.html">Cancellation Policy</a> · <a href="/compliance.html">Compliance</a></p></div><div><h2 class="footer-heading">Contact</h2><p><a href="mailto:info@project2pixel.co.za">info@project2pixel.co.za</a></p><p>Centurion, Gauteng, South Africa</p><p>Supporting clients remotely.</p></div></div><p class="footer-meta">&copy; <span id="currentYear">2026</span> Project<span class="highlight-2">2</span>Pixel (PTY) Ltd. Reg: 2025/625844/07.</p></footer><script src="/script.js"></script>
+<footer><div class="footer-grid"><div><img src="/project2pixel-logo-480.png" alt="Project2Pixel Logo" class="footer-logo" width="480" height="480"><p>Project2Pixel provides practical business support and digital solutions for growing organisations.</p></div><div><h2 class="footer-heading">Navigate</h2><p><a href="/">Home</a> · <a href="/services/">Services</a> · <a href="/about/">About</a> · <a href="/contact/">Contact</a></p><p><a href="/privacy.html">Privacy Policy</a> · <a href="/terms.html">Terms of Service</a> · <a href="/cancellation.html">Cancellation Policy</a> · <a href="/compliance.html">Compliance</a></p></div><div><h2 class="footer-heading">Contact</h2><p><a href="mailto:info@project2pixel.co.za">info@project2pixel.co.za</a></p><p>Centurion, Gauteng, South Africa</p><p>Supporting clients remotely.</p></div></div><p class="footer-meta">&copy; <span id="currentYear">2026</span> Project<span class="highlight-2">2</span>Pixel (PTY) Ltd. Reg: 2025/625844/07.</p></footer><script src="/script.js"></script><script src="/contact-form.js"></script>
 </body></html>
