+++ script.js
@@ -94,58 +94,3 @@ const currentYear = document.getElementById("currentYear")
 if (currentYear) {
   currentYear.textContent = String(new Date().getFullYear())
 }
-
-const enquiryForm = document.querySelector("[data-enquiry-form]")
-
-if (enquiryForm) {
-  const status = enquiryForm.querySelector("[data-form-status]")
-  const submitButton = enquiryForm.querySelector('button[type="submit"]')
-  const serviceSelect = enquiryForm.querySelector("#service")
-  const serviceMap = {
-    "business-support": "Executive & Business Support",
-    websites: "Websites & Digital Presence",
-    "crm-systems": "CRM & Business Systems",
-    "workflow-automation": "Workflow & Automation",
-    "digital-care": "Website & Digital Care",
-  }
-  const requestedService = new URLSearchParams(window.location.search).get("service")
-  if (serviceMap[requestedService]) serviceSelect.value = serviceMap[requestedService]
-
-  enquiryForm.addEventListener("submit", async (event) => {
-    event.preventDefault()
-    if (!enquiryForm.reportValidity()) return
-
-    submitButton.disabled = true
-    status.className = "form-status"
-    status.textContent = "Sending your enquiry…"
-
-    const deliveryUrl = new URL(enquiryForm.action)
-    deliveryUrl.pathname = `/ajax${deliveryUrl.pathname}`
-    const controller = new AbortController()
-    const timeout = window.setTimeout(() => controller.abort(), 15000)
-
-    try {
-      const response = await fetch(deliveryUrl, {
-        method: "POST",
-        headers: { Accept: "application/json" },
-        body: new FormData(enquiryForm),
-        signal: controller.signal,
-      })
-      const result = await response.json().catch(() => ({}))
-      const wasDelivered = result.success === true || result.success === "true"
-      if (!response.ok || !wasDelivered) throw new Error("Delivery failed")
-
-      enquiryForm.reset()
-      status.className = "form-status success"
-      status.textContent = "Thank you. Your enquiry has been received and we'll be in touch."
-      status.focus()
-    } catch {
-      status.className = "form-status error"
-      status.textContent = "We couldn't send your enquiry right now. Please contact us directly by email instead."
-      status.focus()
-    } finally {
-      window.clearTimeout(timeout)
-      submitButton.disabled = false
-    }
-  })
-}
